chore: 🔧 refactor parser code
