+++ src/utils/index.ts
@@ -1,4 +1,4 @@
 export * from "./date";
 export * from "./emojis";
-export * from "./slack_text_to_jsx";
+export * from "./markdown_parser";
 export * from "./is_accessory_stacked";
